Split CPP headers.
Main include file now: <roxx/server.h>.
