+++ demo/cmake/FindRollout.cmake
@@ -11,6 +11,7 @@ else ()
 endif ()
 
 FIND_PATH(ROLLOUT_INCLUDE_DIR rox/server.h)
+FIND_PATH(ROLLOUT_INCLUDE_DIR2 roxx/server.hpp)
 FIND_LIBRARY(ROLLOUT_LIBRARY NAMES rollout)
 SET(ROLLOUT_TARGET_NAME ROLLOUT::SDK)
 
@@ -26,7 +27,7 @@ if (ROLLOUT_FOUND)
     if (NOT TARGET ${ROLLOUT_TARGET_NAME})
         add_library(${ROLLOUT_TARGET_NAME} UNKNOWN IMPORTED)
         set_target_properties(${ROLLOUT_TARGET_NAME} PROPERTIES
-                INTERFACE_INCLUDE_DIRECTORIES "${ROLLOUT_INCLUDE_DIR}"
+                INTERFACE_INCLUDE_DIRECTORIES "${ROLLOUT_INCLUDE_DIR} ${ROLLOUT_INCLUDE_DIR2}"
                 IMPORTED_LOCATION "${ROLLOUT_LIBRARY}")
     endif ()
 else (ROLLOUT_FOUND)
