+++ include/roxx/impression.h
@@ -1,0 +1,17 @@
+#pragma once
+
+extern "C" {
+#include <rox/impression.h>
+}
+
+namespace Rox {
+
+    typedef struct RoxReportingValue ReportingValue;
+
+    class ROX_API ImpressionHandlerInterface {
+    public:
+        virtual void HandleImpression(ReportingValue *value, Context *context) = 0;
+
+        virtual ~ImpressionHandlerInterface() = default;
+    };
+}
