+++ include/roxx/values.h
@@ -1,0 +1,12 @@
+#pragma once
+
+extern "C" {
+#include <rox/values.h>
+}
+
+namespace Rox {
+
+    typedef struct RoxDynamicValue DynamicValue;
+
+
+}
