+++ src/core/properties.c
@@ -1,5 +1,6 @@
 #include <assert.h>
 #include <stdio.h>
+#include <stdlib.h>
 #include "properties.h"
 #include "util.h"
