+++ src/roxx/extensions.c
@@ -1,5 +1,6 @@
 #include <assert.h>
 #include <math.h>
+#include <stdlib.h>
 #include "extensions.h"
 #include "util.h"
