+++ src/values.c
@@ -3,6 +3,7 @@
 #include <float.h>
 #include <stdbool.h>
 #include <string.h>
+#include <stdlib.h>
 
 #include "rox/server.h"
 #include "util.h"
